tests/fdatasync.c: print syscall() result using %ld format

+++ tests/fdatasync.c
@@ -11,10 +11,9 @@ int
 main(void)
 {
 	const long int fd = (long int) 0xdeadbeefffffffff;
-	int rc = syscall(__NR_fdatasync, fd);
-	printf("fdatasync(%d) = %d %s (%m)\n",
-	       (int) fd, rc,
-	       errno2name());
+
+	long rc = syscall(__NR_fdatasync, fd);
+	printf("fdatasync(%d) = %ld %s (%m)\n", (int) fd, rc, errno2name());
 
 	puts("+++ exited with 0 +++");
 	return 0;
